# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Documentation for Furniture App API",
        "description": "Документация API для приложения Furniture App",
        "contact": {
            "email": "support@furnitureapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1/",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/materials": {
            "get": {
                "tags": [
                    "Материалы"
                ],
                "summary": "Получить все материалы",
                "description": "Возвращает список всех материалов.",
                "operationId": "0821961cdcf7173ea92453089dd0a172",
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Material"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Материалы"
                ],
                "summary": "Создать новый материал",
                "description": "Создаёт новый материал и возвращает его.",
                "operationId": "6cd77951b1448e367c3069ed6687379c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Material"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Материал создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ввод"
                    }
                }
            }
        },
        "/api/materials/{id}": {
            "get": {
                "tags": [
                    "Материалы"
                ],
                "summary": "Получить материал по ID",
                "description": "Возвращает материал по его ID.",
                "operationId": "6c25d42a583c7f961ccf8548c969d309",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Материал не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Материалы"
                ],
                "summary": "Обновить материал",
                "description": "Обновляет существующий материал по его ID.",
                "operationId": "182f62eb940ca7d446499048c5ef7047",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Material"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Материал обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Материал не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Материалы"
                ],
                "summary": "Удалить материал",
                "description": "Удаляет существующий материал по его ID.",
                "operationId": "0fa788a436a67a78b74e4da4e847c697",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Материал удалён"
                    },
                    "404": {
                        "description": "Материал не найден"
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Получить все заказы",
                "description": "Возвращает список всех заказов.",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Создать новый заказ",
                "description": "Создаёт новый заказ и возвращает его.",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ввод"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Получить заказ по ID",
                "description": "Возвращает заказ по его ID.",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Обновить заказ",
                "description": "Обновляет существующий заказ по его ID.",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Заказ обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Заказы"
                ],
                "summary": "Удалить заказ",
                "description": "Удаляет существующий заказ по его ID.",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Заказ удалён"
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Material": {
                "title": "Material",
                "description": "Material model",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the material",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Name of the material",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the material",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Order": {
                "required": [
                    "id",
                    "user_id",
                    "total_amount"
                ],
                "properties": {
                    "id": {
                        "description": "ID заказа",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "total_amount": {
                        "description": "Общая сумма заказа",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}